Changed how the server gets the drawn graph to provide to the user.
Instead of accessing the ArgumentGraph instance in main thread from
one of the service-providing threads.  The service providing threads
create their own instances, pulling the most recently committed snapshot
of the graph directly from the database.  This means an incomplete
(because not yet committed to the database) version of the graph will
not be provided to the user.

In the previous version an incomplete version could not be provided
to the user only because the prepare_graph function wasn't called until
the graph was ready to be drawn, and the function to get the graph
used double buffering to only present the most recent completed graph.

This change only improves the code, therefore, it doesn't change the
user-visible behavior of the software.  It is an improvement because the
implementation of drawing the graph does not have to be dependent on
the implementation details of getting the recent graph for use.

Locks are no longer needed in the ArgumentGraph class because concurrency
is handled by the DBMS (ZODB).